Implemented Bounce & Healthup

+++ src/main/java/com/warven22/sodapop/utils/PotionEffectUtil.java
@@ -5,7 +5,7 @@ import net.minecraft.potion.Effect;
 import net.minecraft.potion.EffectInstance;
 
 public class PotionEffectUtil {
-	public static void addEffectToLivingEntity(Effect effect, int durationInSeconds, int level, LivingEntity entity) {
+	public static void addEffectToLivingEntity(Effect effect, float durationInSeconds, int level, LivingEntity entity) {
 		entity.addPotionEffect(new EffectInstance(effect, TimeUtil.getTicksFromSeconds(durationInSeconds), level-1));
 	}
 }
