+++ src/main/java/com/warven22/sodapop/utils/TimeUtil.java
@@ -3,7 +3,7 @@ package com.warven22.sodapop.utils;
 public class TimeUtil {
 	private static int TICKS_PER_SECOND = 20;
 	
-	public static int getTicksFromSeconds(int seconds) {
-		return TICKS_PER_SECOND * seconds;
+	public static int getTicksFromSeconds(float seconds) {
+		return (int) (TICKS_PER_SECOND * seconds);
 	}
 }
